backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
rasterio
fiona
shapely
fastkml
pyyaml
aiofiles
psutil
Pillow
numpy
sqlalchemy
psycopg2-binary
geoalchemy2
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
email-validator
lxml
pyproj
